changed doorbashEaseMultiplier from int to float

+++ Assets/Scripts/SettingsManager.cs
@@ -218,7 +218,7 @@ namespace DaggerfallWorkshop
         public float ColorBoostRadius { get; set; }
         public float ColorBoostIntensity { get; set; }
         public bool LogHudMessages { get; set; }
-        public int DoorBashEaseMultiplier { get; set; }
+        public float DoorBashEaseMultiplier { get; set; }
         public float ColorBoostDungeonScale { get; set; }
         public float ColorBoostExteriorScale { get; set; }
         public float ColorBoostInteriorScale { get; set; }
@@ -523,7 +523,7 @@ namespace DaggerfallWorkshop
             EnableVitalsIndicators = GetBool(sectionGUI, "EnableVitalsIndicators");
             DisplayHUDScaleAdjust = GetFloat(sectionGUI, "DisplayHUDScaleAdjust");
             LogHudMessages = GetBool(sectionGUI, "LogHudMessages");
-            DoorBashEaseMultiplier = GetInt(sectionGUI, "DoorBashEaseMultiplier");
+            DoorBashEaseMultiplier = GetFloat(sectionGUI, "DoorBashEaseMultiplier");
 
             DisplayMobileEnemyHealthStatus = GetBool(sectionGUI, "DisplayMobileEnemyHealthStatus");
 
@@ -757,7 +757,7 @@ namespace DaggerfallWorkshop
             SetBool(sectionGUI, "HideLoginName", HideLoginName);
             SetFloat(sectionGUI, "DisplayHUDScaleAdjust", DisplayHUDScaleAdjust);
             SetBool(sectionGUI, "LogHudMessages", LogHudMessages);
-            SetInt(sectionGUI, "DoorBashEaseMultiplier", DoorBashEaseMultiplier);
+            SetFloat(sectionGUI, "DoorBashEaseMultiplier", DoorBashEaseMultiplier);
             SetBool(sectionGUI, "DisplayMobileEnemyHealthStatus", DisplayMobileEnemyHealthStatus);
 
             SetInt(sectionMeleeAttacks, "MeleeAttackDetection", MeleeAttackDetection);
